+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 - imprint and privacy policy page, editable via settings
 - flag for public poll statistic, if checked poll stats can be shown to everyone with link
+### Changed
+- email in member table is now unique (validation in member form added)
 
 ## v0.7.10 - 2024-05-05
 ### Added